Membenarkan Problem

+++ capstone_flutter/lib/view/screen/pulsa&paket_data_screen/paket_data/pin_paket_data_screen.dart
@@ -84,7 +84,7 @@ class _PinPaketDataScreenState extends State<PinPaketDataScreen> {
       final newBalance =
           widget.balanceNow; // Ganti dengan nilai balance yang baru
       userProvider.updateUserInfo(
-          userProvider.name, userProvider.phone, newBalance as double);
+          userProvider.name, userProvider.phone, newBalance);
       // ignore: use_build_context_synchronously
       Navigator.push(
         context,
