+++ capstone_flutter/lib/view/screen/pulsa&paket_data_screen/pin_pulsa_screen.dart
@@ -85,7 +85,7 @@ class _PinPulsaScreenState extends State<PinPulsaScreen> {
       final newBalance =
           widget.balanceNow; // Ganti dengan nilai balance yang baru
       userProvider.updateUserInfo(
-          userProvider.name, userProvider.phone, newBalance as double);
+          userProvider.name, userProvider.phone, newBalance);
       // ignore: use_build_context_synchronously
       Navigator.push(
         context,
